split and refactor teh jobsController
